Loading Files
Function to load files from a dialog box has been added and all selected files are displayed in the treeview.
Upon adding files to the treeview, the search and rename buttons are enabled. The gui has also been updated to allow
users to enter the name or tvdb id of the show for searching.