Workflow: Test Automation (#6)
* Base layout

Created the window and base structure of the application. Gui widgets can be loaded from a json file dynamically when given the proper inputs.

* Completion of GUI

All required widgets have been created. The GUI engine places and assign all attributes based on the informaiton from the json file.
Functions can be added via the self.functions dict that is passed to the Parser.

* DocStrings and Tests

Created docstrings to explain functionality and most test have been created for the gui (except one for the application menu).

* Updated README

Added new images for read me, created a requirements.txt and filled out
the readme with what information is possible at the current moment.